% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulate_utils.R
\name{check_and_ignore_wrong_type}
\alias{check_and_ignore_wrong_type}
\title{check_and_ignore_wrong_type}
\usage{
check_and_ignore_wrong_type(data, variables, warning_msg, stop_msg)
}
\arguments{
\item{data}{["chr"] Some desc.}

\item{variables}{["chr"] Some desc.}

\item{warning_msg}{["chr"] Some desc.}

\item{stop_msg}{["chr"] Some desc.}
}
\value{
message
}
\description{
check_and_ignore_wrong_type
}
