% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{formatting}
\alias{formatting}
\alias{format_background_color}
\alias{format_border}
\alias{format_border_background_color}
\alias{format_border_bottom}
\alias{format_border_bottom_background_color}
\alias{format_border_bottom_color}
\alias{format_border_color}
\alias{format_border_left}
\alias{format_border_left_background_color}
\alias{format_border_left_color}
\alias{format_border_right}
\alias{format_border_right_background_color}
\alias{format_border_right_color}
\alias{format_border_top}
\alias{format_border_top_background_color}
\alias{format_border_top_color}
\alias{format_color}
\alias{format_column_separator}
\alias{format_column_separator_background_color}
\alias{format_column_separator_color}
\alias{format_corner}
\alias{format_corner_background_color}
\alias{format_corner_bottom_left}
\alias{format_corner_bottom_left_background_color}
\alias{format_corner_bottom_left_color}
\alias{format_corner_bottom_right}
\alias{format_corner_bottom_right_background_color}
\alias{format_corner_bottom_right_color}
\alias{format_corner_color}
\alias{format_corner_top_left}
\alias{format_corner_top_left_background_color}
\alias{format_corner_top_left_color}
\alias{format_corner_top_right}
\alias{format_corner_top_right_background_color}
\alias{format_corner_top_right_color}
\alias{format_font_align}
\alias{format_font_background_color}
\alias{format_font_color}
\alias{format_font_style}
\alias{format_height}
\alias{format_hide_border}
\alias{format_hide_border_bottom}
\alias{format_hide_border_left}
\alias{format_hide_border_right}
\alias{format_hide_border_top}
\alias{format_locale}
\alias{format_multi_byte_characters}
\alias{format_padding}
\alias{format_padding_bottom}
\alias{format_padding_left}
\alias{format_padding_right}
\alias{format_padding_top}
\alias{format_show_border}
\alias{format_show_border_bottom}
\alias{format_show_border_left}
\alias{format_show_border_right}
\alias{format_show_border_top}
\alias{format_width}
\title{Formatting of tables, columns, rows and cells}
\usage{
format_background_color(x, value)

format_border(x, value)

format_border_background_color(x, value)

format_border_bottom(x, value)

format_border_bottom_background_color(x, value)

format_border_bottom_color(x, value)

format_border_color(x, value)

format_border_left(x, value)

format_border_left_background_color(x, value)

format_border_left_color(x, value)

format_border_right(x, value)

format_border_right_background_color(x, value)

format_border_right_color(x, value)

format_border_top(x, value)

format_border_top_background_color(x, value)

format_border_top_color(x, value)

format_color(x, value)

format_column_separator(x, value)

format_column_separator_background_color(x, value)

format_column_separator_color(x, value)

format_corner(x, value)

format_corner_background_color(x, value)

format_corner_bottom_left(x, value)

format_corner_bottom_left_background_color(x, value)

format_corner_bottom_left_color(x, value)

format_corner_bottom_right(x, value)

format_corner_bottom_right_background_color(x, value)

format_corner_bottom_right_color(x, value)

format_corner_color(x, value)

format_corner_top_left(x, value)

format_corner_top_left_background_color(x, value)

format_corner_top_left_color(x, value)

format_corner_top_right(x, value)

format_corner_top_right_background_color(x, value)

format_corner_top_right_color(x, value)

format_font_align(x, value)

format_font_background_color(x, value)

format_font_color(x, value)

format_font_style(x, value)

format_height(x, value)

format_hide_border(x)

format_hide_border_bottom(x)

format_hide_border_left(x)

format_hide_border_right(x)

format_hide_border_top(x)

format_locale(x, value)

format_multi_byte_characters(x, value)

format_padding(x, value)

format_padding_bottom(x, value)

format_padding_left(x, value)

format_padding_right(x, value)

format_padding_top(x, value)

format_show_border(x)

format_show_border_bottom(x)

format_show_border_left(x)

format_show_border_right(x)

format_show_border_top(x)

format_width(x, value)
}
\arguments{
\item{x}{A table, column, row or cell.}

\item{value}{The value assumed by the selected parameter.}
}
\value{
The same as the input invisibly.
}
\description{
Allows changing display characteristics of tables, columns, rows and cells.
}
\examples{
table <- tabulate_table()
table \%>\%
  table_add_row("hello") \%>\%
  table_add_row("world")

table \%>\%
  format_hide_border_bottom() \%>\%
  format_width(50)

table

}
