% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulate_utils.R
\name{gather_longer}
\alias{gather_longer}
\title{gather_longer}
\usage{
gather_longer(
  data,
  variables,
  .names_to,
  .values_to,
  .keep_variables = NULL,
  .values_drop_na = FALSE
)
}
\arguments{
\item{data}{[tibble] Some desc.}

\item{variables}{[chr] Some desc.}

\item{.names_to}{[chr] Some desc.}

\item{.values_to}{[chr] Some desc.}

\item{.keep_variables}{[chr] Some desc.}

\item{.values_drop_na}{[chr] Some desc.}
}
\value{
[tibble]
}
\description{
gather_longer
}
