% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulate.R
\name{tabulate}
\alias{tabulate}
\title{tabulate}
\usage{
tabulate(
  data,
  cols,
  weights = NULL,
  groups = NULL,
  samples = NULL,
  return_mean = FALSE,
  values_drop_na = TRUE,
  variable_sep = NULL,
  variable_sep_suffix = NULL,
  keep_empty_levels = FALSE
)
}
\arguments{
\item{data}{A dataframe}

\item{cols}{[str] string vector of variables or tidy supported selection ex. matches(...)}

\item{weights}{[str] string vector of variables or tidy supported selection ex. matches(...)}

\item{groups}{[str] string vector of variables or tidy supported selection ex. matches(...)}

\item{samples}{[str] string vector of variables or tidy supported selection ex. matches(...)}

\item{return_mean}{[boolean] for nominal variables}

\item{values_drop_na}{[boolean] if the base and pct should include NAs}

\item{variable_sep}{[str] regex used to separate the variable column}

\item{variable_sep_suffix}{[str] name of the variable_sep right-hand side output column}

\item{keep_empty_levels}{[str] no levels imputations}
}
\value{
frequency table, dataframe/tibble,   A standardized ET
}
\description{
tji asd asd sda sad sa d tji asd asd sda sad sa  tji asd asd sda sad sa  tji asd asd sda sad sa
tji asd asd sda sad sa  tji asd asd sda sad sa tji asd asd sda sad sa  tji asd asd sda sad sa
#'
}
\examples{

tabulate(iris, Species)

}
