% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo.R
\name{tabulate_demo}
\alias{tabulate_demo}
\title{Generated demonstration tables}
\usage{
tabulate_demo()
}
\value{
\code{NULL} invisibly.
}
\description{
Print two tables showing most of possible functionalities in tabulate.
}
