% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{tabulate_table}
\alias{tabulate_table}
\alias{table_add_row}
\title{Create a new tabulate table}
\usage{
tabulate_table()

table_add_row(table, row)
}
\arguments{
\item{row}{A character vector or a tabulate table.}
}
\description{
Create a new tabulate table
}
\section{Functions}{
\itemize{
\item \code{table_add_row}: Adds rows to the table.
}}

\examples{
table <- tabulate_table() \%>\%
  table_add_row("hello") \%>\%
  table_add_row("world")
table
}
